Split ShellContextMenu and fix submenus
