+++ src/ManagedShell.ShellFolders/ShellNewMenuCommand.cs
@@ -9,30 +9,38 @@ namespace ManagedShell.ShellFolders
 {
     public class ShellNewMenuCommand : ShellMenuCommand
     {
-        public NativeContextMenu AddSubMenu(ShellFolder folder, int position, ref IntPtr parentNativeMenuPtr)
+        internal IContextMenu iContextMenu;
+        internal IContextMenu2 iContextMenu2;
+        internal IContextMenu3 iContextMenu3;
+
+        internal IntPtr iContextMenuPtr;
+        internal IntPtr iContextMenu2Ptr;
+        internal IntPtr iContextMenu3Ptr;
+
+        internal IntPtr nativeMenuPtr;
+        
+        public void AddSubMenu(ShellFolder folder, int position, ref IntPtr parentNativeMenuPtr)
         {
-            NativeContextMenu menu = new NativeContextMenu();
-            
-            if (GetNewContextMenu(folder, ref menu))
+            if (GetNewContextMenu(folder))
             {
-                menu.iContextMenu.QueryContextMenu(
+                iContextMenu.QueryContextMenu(
                     parentNativeMenuPtr,
                     (uint)position,
                     Interop.CMD_FIRST,
                     Interop.CMD_LAST,
                     CMF.NORMAL);
 
-                menu.nativeMenuPtr = Interop.GetSubMenu(parentNativeMenuPtr, position);
+                nativeMenuPtr = Interop.GetSubMenu(parentNativeMenuPtr, position);
 
-                if (Marshal.QueryInterface(menu.iContextMenuPtr, ref Interop.IID_IContextMenu2,
-                    out menu.iContextMenu2Ptr) == NativeMethods.S_OK)
+                if (Marshal.QueryInterface(iContextMenuPtr, ref Interop.IID_IContextMenu2,
+                    out iContextMenu2Ptr) == NativeMethods.S_OK)
                 {
-                    if (menu.iContextMenu2Ptr != IntPtr.Zero)
+                    if (iContextMenu2Ptr != IntPtr.Zero)
                     {
                         try
                         {
-                            menu.iContextMenu2 =
-                                (IContextMenu2)Marshal.GetTypedObjectForIUnknown(menu.iContextMenu2Ptr, typeof(IContextMenu2));
+                            iContextMenu2 =
+                                (IContextMenu2)Marshal.GetTypedObjectForIUnknown(iContextMenu2Ptr, typeof(IContextMenu2));
                         }
                         catch (Exception e)
                         {
@@ -41,15 +49,15 @@ namespace ManagedShell.ShellFolders
                     }
                 }
 
-                if (Marshal.QueryInterface(menu.iContextMenuPtr, ref Interop.IID_IContextMenu3,
-                    out menu.iContextMenu3Ptr) == NativeMethods.S_OK)
+                if (Marshal.QueryInterface(iContextMenuPtr, ref Interop.IID_IContextMenu3,
+                    out iContextMenu3Ptr) == NativeMethods.S_OK)
                 {
-                    if (menu.iContextMenu3Ptr != IntPtr.Zero)
+                    if (iContextMenu3Ptr != IntPtr.Zero)
                     {
                         try
                         {
-                            menu.iContextMenu3 =
-                                (IContextMenu3)Marshal.GetTypedObjectForIUnknown(menu.iContextMenu3Ptr, typeof(IContextMenu3));
+                            iContextMenu3 =
+                                (IContextMenu3)Marshal.GetTypedObjectForIUnknown(iContextMenu3Ptr, typeof(IContextMenu3));
                         }
                         catch (Exception e)
                         {
@@ -58,23 +66,21 @@ namespace ManagedShell.ShellFolders
                     }
                 }
             }
-
-            return menu;
         }
 
-        private bool GetNewContextMenu(ShellFolder folder, ref NativeContextMenu menu)
+        private bool GetNewContextMenu(ShellFolder folder)
         {
             if (Interop.CoCreateInstance(
                 ref Interop.CLSID_NewMenu,
                 IntPtr.Zero,
                 CLSCTX.INPROC_SERVER,
                 ref Interop.IID_IContextMenu,
-                out menu.iContextMenuPtr) == NativeMethods.S_OK)
+                out iContextMenuPtr) == NativeMethods.S_OK)
             {
-                menu.iContextMenu = Marshal.GetTypedObjectForIUnknown(menu.iContextMenuPtr, typeof(IContextMenu)) as IContextMenu;
+                iContextMenu = Marshal.GetTypedObjectForIUnknown(iContextMenuPtr, typeof(IContextMenu)) as IContextMenu;
 
                 if (Marshal.QueryInterface(
-                    menu.iContextMenuPtr,
+                    iContextMenuPtr,
                     ref Interop.IID_IShellExtInit,
                     out var iShellExtInitPtr) == NativeMethods.S_OK)
                 {
@@ -98,6 +104,51 @@ namespace ManagedShell.ShellFolders
             }
             
             return false;
+        }
+
+        internal void FreeResources()
+        {
+            if (iContextMenu != null)
+            {
+                Marshal.FinalReleaseComObject(iContextMenu);
+                iContextMenu = null;
+            }
+
+            if (iContextMenu2 != null)
+            {
+                Marshal.FinalReleaseComObject(iContextMenu2);
+                iContextMenu2 = null;
+            }
+
+            if (iContextMenu3 != null)
+            {
+                Marshal.FinalReleaseComObject(iContextMenu3);
+                iContextMenu3 = null;
+            }
+
+            if (iContextMenuPtr != IntPtr.Zero)
+            {
+                Marshal.Release(iContextMenuPtr);
+                iContextMenuPtr = IntPtr.Zero;
+            }
+
+            if (iContextMenu2Ptr != IntPtr.Zero)
+            {
+                Marshal.Release(iContextMenu2Ptr);
+                iContextMenu2Ptr = IntPtr.Zero;
+            }
+
+            if (iContextMenu3Ptr != IntPtr.Zero)
+            {
+                Marshal.Release(iContextMenu3Ptr);
+                iContextMenu3Ptr = IntPtr.Zero;
+            }
+
+            if (nativeMenuPtr != IntPtr.Zero)
+            {
+                Interop.DestroyMenu(nativeMenuPtr);
+                nativeMenuPtr = IntPtr.Zero;
+            }
         }
     }
 }
